move BottomPromo content
